feat(dolos): add testgen to build utils, fix docker
