+++ crates/build_utils/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2021"
 
 [dependencies]
 toml = "0.8"
+bzip2 = "0.4.3"
+tar = "0.4.38"
+zip = "0.6.6"
+reqwest = { version = "0.12.7", features = ["blocking"] }
 
 [lints.clippy]
 uninlined_format_args = "deny"
