+++ crates/build_utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod dolos;
 pub mod git;
+pub mod testgen_hs;
